+++ R/DatabaseConnector.R
@@ -31,8 +31,13 @@ NULL
 #' @name jdbcDrivers
 #'
 #' @title
-#' How to download JDBC drivers for the various data platforms.
-#'
+#' How to download and use JDBC drivers for the various data platforms.
+#' 
+#' @description 
+#' Below are instructions for downloading JDBC drivers for the various data platforms. Once downloaded
+#' use the \code{pathToDriver} argument in the \code{\link{connect}} or \code{\link{createConnectionDetails}}
+#' functions to point to the driver.
+#' 
 #' @section
 #' PostgresSql: Go to \href{https://jdbc.postgresql.org/download.html}{the PostgresSQL JDBC site} and
 #' download the current version. The file is called something like 'postgresql-42.2.2.jar'.
@@ -68,6 +73,7 @@ NULL
 #' Impala: Go to
 #' \href{https://www.cloudera.com/downloads/connectors/impala/jdbc/2-5-5.html}{Cloudera's site}, pick
 #' your OS version, and click "GET IT NOW!'. Register, and you should be able to download the driver.
+#' 
 NULL
 
 # Borrowed from devtools: https://github.com/hadley/devtools/blob/ba7a5a4abd8258c52cb156e7b26bb4bf47a79f0b/R/utils.r#L44