+++ R/Sql.R
@@ -1,6 +1,6 @@
 # @file Sql.R
 #
-# Copyright 2018 Observational Health Data Sciences and Informatics
+# Copyright 2019 Observational Health Data Sciences and Informatics
 #
 # This file is part of DatabaseConnector
 #
@@ -45,12 +45,12 @@
              sep = "\n"), call. = FALSE)
 }
 
-as.POSIXct.ff_vector <- function(x) {
+as.POSIXct.ff_vector <- function(x, ...) {
   chunks <- bit::chunk(x)
   i <- chunks[[1]]
-  result <- ff::as.ff(as.POSIXct(x[i]))
+  result <- ff::as.ff(as.POSIXct(x[i], ...))
   for (i in chunks[-1]) {
-    result <- ffbase::ffappend(result, ff::as.ff(as.POSIXct(x[i])))
+    result <- ffbase::ffappend(result, ff::as.ff(as.POSIXct(x[i], ...)))
   }
   return(result)
 }
@@ -77,6 +77,11 @@ as.POSIXct.ff_vector <- function(x) {
 #'
 #' @export
 lowLevelQuerySql.ffdf <- function(connection, query = "", datesAsString = FALSE) {
+  UseMethod("lowLevelQuerySql.ffdf", connection)
+}
+
+#' @export
+lowLevelQuerySql.ffdf.default <- function(connection, query = "", datesAsString = FALSE) {
   if (rJava::is.jnull(connection@jConnection))
     stop("Connection is closed")
   batchedQuery <- rJava::.jnew("org.ohdsi.databaseConnector.BatchedQuery",
@@ -137,6 +142,21 @@ lowLevelQuerySql.ffdf <- function(connection, query = "", datesAsString = FALSE)
   return(ffdf)
 }
 
+#' @export
+lowLevelQuerySql.ffdf.DatabaseConnectorDbiConnection <- function(connection, query = "", datesAsString = FALSE) {
+  results <- lowLevelQuerySql(connection, query)
+  for (i in 1:ncol(results)) {
+    if (class(results[, i]) == "character") {
+      results[, i] <- as.factor(results[, i])
+    }
+  }
+  if (nrow(results) == 0) {
+    return(results)
+  } else {
+    return(ff::as.ffdf(results))
+  }
+}
+
 #' Low level function for retrieving data to a data frame
 #'
 #' @description
@@ -157,6 +177,18 @@ lowLevelQuerySql.ffdf <- function(connection, query = "", datesAsString = FALSE)
 #'
 #' @export
 lowLevelQuerySql <- function(connection, query = "", datesAsString = FALSE) {
+  # Not using UseMethod pattern to avoid note about lowLevelQuerySql.ffdf being a method:
+  if (inherits(connection, "DatabaseConnectorDbiConnection")) {
+    results <- DBI::dbGetQuery(connection@dbiConnection, query)
+    # For compatibility with JDBC:
+    for (i in 1:ncol(results)) {
+      if (class(results[, i]) == "integer64") {
+        results[, i] <- as.numeric(results[, i])
+      }
+    }
+    return(results)
+  }
+  
   if (rJava::is.jnull(connection@jConnection))
     stop("Connection is closed")
   batchedQuery <- rJava::.jnew("org.ohdsi.databaseConnector.BatchedQuery",
@@ -209,9 +241,24 @@ lowLevelQuerySql <- function(connection, query = "", datesAsString = FALSE) {
 #'
 #' @export
 lowLevelExecuteSql <- function(connection, sql) {
+  UseMethod("lowLevelExecuteSql", connection)
+}
+
+#' @export
+lowLevelExecuteSql.default <- function(connection, sql) {
   statement <- rJava::.jcall(connection@jConnection, "Ljava/sql/Statement;", "createStatement")
   on.exit(rJava::.jcall(statement, "V", "close"))
-  rowsAffected <- rJava::.jcall(statement, "I", "executeUpdate", as.character(sql), check = FALSE)
+  hasResultSet <- rJava::.jcall(statement, "Z", "execute", as.character(sql), check = FALSE)
+  rowsAffected <- 0
+  if (!hasResultSet) {
+    rowsAffected <- rJava::.jcall(statement, "I", "getUpdateCount", check = FALSE)
+  }
+  invisible(rowsAffected)
+}
+
+#' @export
+lowLevelExecuteSql.DatabaseConnectorDbiConnection <- function(connection, sql) {
+  rowsAffected <- DBI::dbExecute(connection@dbiConnection, sql)
   invisible(rowsAffected)
 }
 
@@ -240,7 +287,7 @@ lowLevelExecuteSql <- function(connection, sql) {
 #'
 #' @examples
 #' \dontrun{
-#' connectionDetails <- createConnectionDetails(dbms = "mysql",
+#' connectionDetails <- createConnectionDetails(dbms = "postgresql",
 #'                                              server = "localhost",
 #'                                              user = "root",
 #'                                              password = "blah",
@@ -256,7 +303,7 @@ executeSql <- function(connection,
                        progressBar = TRUE,
                        reportOverallTime = TRUE, 
                        errorReportFile = file.path(getwd(), "errorReport.txt")) {
-  if (rJava::is.jnull(connection@jConnection))
+  if (inherits(connection, "DatabaseConnectorJdbcConnection") && rJava::is.jnull(connection@jConnection))
     stop("Connection is closed")
   if (profile)
     progressBar <- FALSE
@@ -284,7 +331,7 @@ executeSql <- function(connection,
     if (progressBar)
       setTxtProgressBar(pb, i/length(sqlStatements))
   }
-  if (!rJava::.jcall(connection@jConnection, "Z", "getAutoCommit")) {
+  if (inherits(connection, "DatabaseConnectorJdbcConnection") && !rJava::.jcall(connection@jConnection, "Z", "getAutoCommit")) {
     rJava::.jcall(connection@jConnection, "V", "commit")
   }
   if (progressBar)
@@ -295,15 +342,42 @@ executeSql <- function(connection,
   }
 }
 
+convertFields <- function(dbms, result) {
+  if (dbms == "impala") {
+    for (colname in colnames(result)) {
+      if (grepl("DATE", colname)) {
+        result[[colname]] <- as.Date(result[[colname]], "%Y-%m-%d")
+      }
+    }
+  }
+  if (dbms == "sqlite") {
+    for (colname in colnames(result)) {
+      if (grepl("DATE$", colname)) {
+        result[[colname]] <- as.Date(as.POSIXct(result[[colname]], origin = "1970-01-01", tz = "GMT"))
+      }
+      if (grepl("DATETIME$", colname)) {
+        if (ff::is.ffdf(result)) {
+          result[[colname]] <- as.POSIXct.ff_vector(result[[colname]], origin = "1970-01-01", tz = "GMT")
+        } else {
+          result[[colname]] <- as.POSIXct(result[[colname]], origin = "1970-01-01", tz = "GMT")
+        }
+      }
+      
+    }
+  } 
+  return(result)
+}
+
 #' Retrieve data to a data.frame
 #'
 #' @description
 #' This function sends SQL to the server, and returns the results.
 #'
-#' @param connection        The connection to the database server.
-#' @param sql               The SQL to be send.
-#' @param errorReportFile   The file where an error report will be written if an error occurs. Defaults to
-#'                          'errorReport.txt' in the current working directory.
+#' @param connection           The connection to the database server.
+#' @param sql                  The SQL to be send.
+#' @param errorReportFile      The file where an error report will be written if an error occurs. Defaults to
+#'                             'errorReport.txt' in the current working directory.
+#' @param snakeCaseToCamelCase If true, field names are assumed to use snake_case, and are converted to camelCase.  
 #'
 #' @details
 #' This function sends the SQL to the server and retrieves the results. If an error occurs during SQL
@@ -315,7 +389,7 @@ executeSql <- function(connection,
 #'
 #' @examples
 #' \dontrun{
-#' connectionDetails <- createConnectionDetails(dbms = "mysql",
+#' connectionDetails <- createConnectionDetails(dbms = "postgresql",
 #'                                              server = "localhost",
 #'                                              user = "root",
 #'                                              password = "blah",
@@ -325,8 +399,8 @@ executeSql <- function(connection,
 #' disconnect(conn)
 #' }
 #' @export
-querySql <- function(connection, sql, errorReportFile = file.path(getwd(), "errorReport.txt")) {
-  if (rJava::is.jnull(connection@jConnection))
+querySql <- function(connection, sql, errorReportFile = file.path(getwd(), "errorReport.txt"), snakeCaseToCamelCase = FALSE) {
+  if (inherits(connection, "DatabaseConnectorJdbcConnection") && rJava::is.jnull(connection@jConnection))
     stop("Connection is closed")
   # Calling splitSql, because this will also strip trailing semicolons (which cause Oracle to crash).
   sqlStatements <- SqlRender::splitSql(sql)
@@ -337,12 +411,9 @@ querySql <- function(connection, sql, errorReportFile = file.path(getwd(), "erro
   tryCatch({
     result <- lowLevelQuerySql(connection, sqlStatements[1])
     colnames(result) <- toupper(colnames(result))
-    if (attr(connection, "dbms") == "impala") {
-      for (colname in colnames(result)) {
-        if (grepl("DATE", colname)) {
-          result[[colname]] <- as.Date(result[[colname]], "%Y-%m-%d")
-        }
-      }
+    result <- convertFields(connection@dbms, result)
+    if (snakeCaseToCamelCase) {
+      colnames(result) <- SqlRender::snakeCaseToCamelCase(colnames(result))
     }
     return(result)
   }, error = function(err) {
@@ -355,10 +426,11 @@ querySql <- function(connection, sql, errorReportFile = file.path(getwd(), "erro
 #' @description
 #' This function sends SQL to the server, and returns the results in an ffdf object.
 #'
-#' @param connection        The connection to the database server.
-#' @param sql               The SQL to be send.
-#' @param errorReportFile   The file where an error report will be written if an error occurs. Defaults to
-#'                          'errorReport.txt' in the current working directory.
+#' @param connection           The connection to the database server.
+#' @param sql                  The SQL to be send.
+#' @param errorReportFile      The file where an error report will be written if an error occurs. Defaults to
+#'                             'errorReport.txt' in the current working directory.
+#' @param snakeCaseToCamelCase If true, field names are assumed to use snake_case, and are converted to camelCase.
 #'
 #' @details
 #' Retrieves data from the database server and stores it in an ffdf object. This allows very large
@@ -373,7 +445,7 @@ querySql <- function(connection, sql, errorReportFile = file.path(getwd(), "erro
 #' @examples
 #' \dontrun{
 #' library(ffbase)
-#' connectionDetails <- createConnectionDetails(dbms = "mysql",
+#' connectionDetails <- createConnectionDetails(dbms = "postgresql",
 #'                                              server = "localhost",
 #'                                              user = "root",
 #'                                              password = "blah",
@@ -383,8 +455,8 @@ querySql <- function(connection, sql, errorReportFile = file.path(getwd(), "erro
 #' disconnect(conn)
 #' }
 #' @export
-querySql.ffdf <- function(connection, sql, errorReportFile = file.path(getwd(), "errorReport.txt")) {
-  if (rJava::is.jnull(connection@jConnection))
+querySql.ffdf <- function(connection, sql, errorReportFile = file.path(getwd(), "errorReport.txt"), snakeCaseToCamelCase = FALSE) {
+  if (inherits(connection, "DatabaseConnectorJdbcConnection") && rJava::is.jnull(connection@jConnection))
     stop("Connection is closed")
   # Calling splitSql, because this will also strip trailing semicolons (which cause Oracle to crash).
   sqlStatements <- SqlRender::splitSql(sql)
@@ -395,15 +467,163 @@ querySql.ffdf <- function(connection, sql, errorReportFile = file.path(getwd(), 
   tryCatch({
     result <- lowLevelQuerySql.ffdf(connection, sqlStatements[1])
     colnames(result) <- toupper(colnames(result))
-    if (attr(connection, "dbms") == "impala") {
-      for (colname in colnames(result)) {
-        if (grepl("DATE", colname)) {
-          result[[colname]] <- as.Date(result[[colname]], "%Y-%m-%d")
-        }
-      }
+    result <- convertFields(connection@dbms, result)
+    if (snakeCaseToCamelCase) { 
+      colnames(result) <- SqlRender::snakeCaseToCamelCase(colnames(result))
     }
     return(result)
   }, error = function(err) {
     .createErrorReport(connection@dbms, err$message, sql, errorReportFile)
   })
+}
+
+#' Render, translate, execute SQL code
+#'
+#' @description
+#' This function renders, translates, and executes SQL consisting of one or more statements.
+#'
+#' @param connection          The connection to the database server.
+#' @param sql                 The SQL to be executed
+#' @param profile             When true, each separate statement is written to file prior to sending to
+#'                            the server, and the time taken to execute a statement is displayed.
+#' @param progressBar         When true, a progress bar is shown based on the statements in the SQL
+#'                            code.
+#' @param reportOverallTime   When true, the function will display the overall time taken to execute
+#'                            all statements.
+#' @param errorReportFile     The file where an error report will be written if an error occurs. Defaults to
+#'                            'errorReport.txt' in the current working directory.
+#' @param oracleTempSchema    A schema that can be used to create temp tables in when using Oracle or Impala.
+#' @param ...                 Parameters that will be used to render the SQL.
+#'
+#' @details
+#' This function calls the \code{render} and \code{translate} functions in the SqlRender package before 
+#' calling \code{\link{executeSql}}.
+#'
+#' @examples
+#' \dontrun{
+#' connectionDetails <- createConnectionDetails(dbms = "postgresql",
+#'                                              server = "localhost",
+#'                                              user = "root",
+#'                                              password = "blah",
+#'                                              schema = "cdm_v4")
+#' conn <- connect(connectionDetails)
+#' renderTranslateExecuteSql(connection, 
+#'                           sql = "SELECT * INTO #temp FROM @@schema.person;",
+#'                           schema = "cdm_synpuf")
+#' disconnect(conn)
+#' }
+#' @export
+renderTranslateExecuteSql <- function(connection,
+                                      sql,
+                                      profile = FALSE,
+                                      progressBar = TRUE,
+                                      reportOverallTime = TRUE, 
+                                      errorReportFile = file.path(getwd(), "errorReport.txt"),
+                                      oracleTempSchema = NULL,
+                                      ...) {
+  sql <- SqlRender::render(sql, ...)
+  sql <- SqlRender::translate(sql, targetDialect = connection@dbms, oracleTempSchema = oracleTempSchema)
+  executeSql(connection = connection,
+             sql = sql,
+             profile = profile,
+             progressBar = progressBar,
+             reportOverallTime = reportOverallTime,
+             errorReportFile = errorReportFile)
+}
+
+#' Render, translate, and query to data.frame
+#'
+#' @description
+#' This function renders, and translates SQL, sends it to the server, and returns the results as a data.frame.
+#'
+#' @param connection           The connection to the database server.
+#' @param sql                  The SQL to be send.
+#' @param errorReportFile      The file where an error report will be written if an error occurs. Defaults to
+#'                             'errorReport.txt' in the current working directory.
+#' @param snakeCaseToCamelCase If true, field names are assumed to use snake_case, and are converted to camelCase.  
+#' @param oracleTempSchema     A schema that can be used to create temp tables in when using Oracle or Impala.
+#' @param ...                  Parameters that will be used to render the SQL.
+#'
+#' @details
+#' This function calls the \code{render} and \code{translate} functions in the SqlRender package before 
+#' calling \code{\link{querySql}}.
+#'
+#' @return
+#' A data frame.
+#'
+#' @examples
+#' \dontrun{
+#' connectionDetails <- createConnectionDetails(dbms = "postgresql",
+#'                                              server = "localhost",
+#'                                              user = "root",
+#'                                              password = "blah",
+#'                                              schema = "cdm_v4")
+#' conn <- connect(connectionDetails)
+#' persons <- renderTranslatequerySql(conn, 
+#'                                    sql = "SELECT TOP 10 * FROM @@schema.person",
+#'                                    schema = "cdm_synpuf")
+#' disconnect(conn)
+#' }
+#' @export
+renderTranslateQuerySql <- function(connection, 
+                                    sql, 
+                                    errorReportFile = file.path(getwd(), "errorReport.txt"), 
+                                    snakeCaseToCamelCase = FALSE,
+                                    oracleTempSchema = NULL,
+                                    ...) {
+  sql <- SqlRender::render(sql, ...)
+  sql <- SqlRender::translate(sql, targetDialect = connection@dbms, oracleTempSchema = oracleTempSchema)
+  return(querySql(connection = connection,
+                  sql = sql,
+                  errorReportFile = errorReportFile,
+                  snakeCaseToCamelCase = snakeCaseToCamelCase))
+}
+
+
+#' Render, translate, and query to ffdf
+#'
+#' @description
+#' This function renders, and translates SQL, sends it to the server, and returns the results as an ffdf object
+#'
+#' @param connection           The connection to the database server.
+#' @param sql                  The SQL to be send.
+#' @param errorReportFile      The file where an error report will be written if an error occurs. Defaults to
+#'                             'errorReport.txt' in the current working directory.
+#' @param snakeCaseToCamelCase If true, field names are assumed to use snake_case, and are converted to camelCase.  
+#' @param oracleTempSchema     A schema that can be used to create temp tables in when using Oracle or Impala.
+#' @param ...                  Parameters that will be used to render the SQL.
+#'
+#' @details
+#' This function calls the \code{render} and \code{translate} functions in the SqlRender package before 
+#' calling \code{\link{querySql.ffdf}}.
+#'
+#' @return
+#' An ffdf object
+#'
+#' @examples
+#' \dontrun{
+#' connectionDetails <- createConnectionDetails(dbms = "postgresql",
+#'                                              server = "localhost",
+#'                                              user = "root",
+#'                                              password = "blah",
+#'                                              schema = "cdm_v4")
+#' conn <- connect(connectionDetails)
+#' persons <- renderTranslatequerySql.ffdf(conn, 
+#'                                         sql = "SELECT * FROM @@schema.person",
+#'                                         schema = "cdm_synpuf")
+#' disconnect(conn)
+#' }
+#' @export
+renderTranslateQuerySql.ffdf <- function(connection, 
+                                         sql, 
+                                         errorReportFile = file.path(getwd(), "errorReport.txt"), 
+                                         snakeCaseToCamelCase = FALSE,
+                                         oracleTempSchema = NULL,
+                                         ...) {
+  sql <- SqlRender::render(sql, ...)
+  sql <- SqlRender::translate(sql, targetDialect = connection@dbms, oracleTempSchema = oracleTempSchema)
+  return(querySql.ffdf(connection = connection,
+                       sql = sql,
+                       errorReportFile = errorReportFile,
+                       snakeCaseToCamelCase = snakeCaseToCamelCase))
 }